move schemas
- Next.js does not like any exports in routes other than route definitions. So move the shared zod schema to a common location.
